First code import
Proof of concept implementation for a web based simulator for html
gaguges in Microsoft Flight Simulator. It consists of a nodejs web server
that serves the HTML/JS/CSS/XML files from the MSFS instalation and
browser javascript mocks for game APIs.

Ttwo panels are currently supported, for Cessna 172 and Bonanza G36,
both with the standard G1000layout (PFD + MFD). A new instrument
(AS1000 Bezel) was implemented to allow interaton with the panel
(button, knobs, etc.)

Basic stuff works, but there is plenty of things that are currently not
working, most notably the map. There are also some scaling issues,
input (letter selector) does not work for flightplans or waipoints and
a few  hacks that need to be addressed.

More panels can be added by updating the CockpitPanelList.xml file with
the location of the instruments.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>