server: fix coui:// patch
We are currently  patching "js/common.js" to fix the hard-coded origin
(coui://). However the patch discarded the origin altogether and this was
causing issues in matching already loaded scripts and causing a lot of js
exceptions like the one below:

Uncaught SyntaxError: Identifier 'RangeDisplay' was already been declared.

To fix this replace coui:/ with window.location.origin since the script
loading checks for the full URL including the origin.

Note that there are still a couple of issues with scripts loading twice due
to inconsistent usage for pulling in scripts, with both <import-script>
and <script> beeing used. To be investigated and fixed later.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>